third_party/WebKit/LayoutTests/fast/viewport/viewport-28-expected.txt

ALERT: viewport size 352x387.2 scale 0.909091 with limits [0.909091, 5] and userScalable true
